Implement basic testing framework with Catch2
The source code for Catch2 v2.13.4 is currently included in the repo.
This is simpler for new users since they do not have to handle git
submodules.

The header file only approach has not been used since it is
unrecommended by the Catch2 documentation.

V2 is used since the documentation is incomplete for V3.